organized header files

+++ src/compiler/struct.c
@@ -1,5 +1,5 @@
 #include "struct.h"
 
-MxcStruct New_MxcStruct(char *name, struct NodeVariable **f, int nf) {
+MxcStruct New_MxcStruct(char *name, struct NodeVariable **f, size_t nf) {
     return (MxcStruct){name, f, nf};
 }
